requirements.txt
Pillow~=8.4.0
Pygments~=2.10.0